backend/requirements.txt
# FastAPI and server
fastapi
uvicorn[standard]
python-multipart
python-dotenv

# Data handling
pandas
numpy<2

# CORS for React frontend
fastapi-cors

# Groq API client
groq

# Vector database
chromadb

# Embeddings
sentence-transformers

# 🔑 CPU-only PyTorch (CRITICAL)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu

# PDF processing
PyPDF2
pdfplumber
python-magic

# Utilities
aiofiles
pydantic
pydantic-settings

# For production
gunicorn
# Observability persistence
sqlalchemy
